backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
cryptography==43.0.1
pyOpenSSL==24.2.1
requests==2.32.3
httpx==0.27.2
jinja2==3.1.4
python-dotenv==1.0.1
rich==13.7.1
dnspython==2.6.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
